Fix Bug in add_form_overview()
When in sub_forms table every subform is in a different mainform, then 'merge_cells' is an empty data frame. The attempt to merge all rows in this empty data frame fails with an error. -> now there is a check for if 'merge_cells' is empty.

Enhance display when certain type of forms is not available in the data set:
Include not available message instead of empty table.